Speedup 'BlockPos', by fixing inlining.
The getters weren't being inlined in simple utilitiy methods like 'World.isValidLocation()'.
This was making these simple methods take longer than they should.

This should result in across the board speedups for any method that accesses blocks or manipulates positions.

See the patch header for an in-depth description of why this was happening.