Revert "Remove some debug checks from BlockStateList"
For 1) you undid your array patch with this
but 2) My patch to this was to remove .containsKey, we always had the null checks. I had that change pulled into Vanilla. Removing null checks is excessive, and will move the NPE to a completely different location in the source code, and makes tracking it down much harder.

Now it is unlikely to ever happen, but still hard to justify the change.