Backport fix for the tick-loop bug by @chickeneer.
This is a backport PaperMC/Paper@ff73ad6dd to the devil version.
Big thanks to @PhanaticD to pointing out the exact fix I needed to make,
so I could do it in a couple minutes.

This is a purely sadistic commit, and I intend to tourment you
with the illusion of progress.

THIS IS NOT AN INDICATOR OF THINGS TO COME,
THOSE WHO USE 1.8.8 HAVE ALREADY SOLD THEIR SOULS TO HEROBRINE.

Fixes #85